+++ app/controllers/doorkeeper/application_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Doorkeeper
   class ApplicationController <
     Doorkeeper.configuration.base_controller.constantize