+++ lib/doorkeeper/errors.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Doorkeeper
   module Errors
     class DoorkeeperError < StandardError
@@ -46,6 +48,7 @@ module Doorkeeper
 
     UnableToGenerateToken = Class.new(DoorkeeperError)
     TokenGeneratorNotFound = Class.new(DoorkeeperError)
+    NoOrmCleaner = Class.new(DoorkeeperError)
 
     InvalidToken = Class.new BaseResponseError
     TokenExpired = Class.new InvalidToken