+++ lib/doorkeeper/grape/helpers.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'doorkeeper/grape/authorization_decorator'
 
 module Doorkeeper