+++ lib/doorkeeper/helpers/controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # Define methods that can be called in any controller that inherits from
 # Doorkeeper::ApplicationMetalController or Doorkeeper::ApplicationController
 module Doorkeeper