+++ NEWS.md
@@ -12,7 +12,8 @@ User-visible changes worth mentioning.
   triggers every time)
 - [#1149] Fix for `URIChecker#valid_for_authorization?` false negative when query is blank, but `?` present.
 - [#1151] Fix Refresh Token strategy: add proper validation of client credentials both for Public & Private clients.
-- [#1152] Fixmigration template: change resource owner data type from integer to references
+- [#1152] Fix migration template: change resource owner data type from integer to Rails generic `references`
+- [#1154] Refactor `StaleRecordsCleaner` to be ORM agnostic.
 
 ## 5.0.0